+++ tests/integration/Serialize.XmlSerialize.test.js
@@ -29,7 +29,7 @@ const {
     XML: { Normalize: { Factory: XmlNormalizeFactory } },
     XmlSerializer
   },
-  Spec: { Spec1dot2, Spec1dot3, Spec1dot4 },
+  Spec: { Spec1dot2, Spec1dot3, Spec1dot4, Spec1dot5 },
   Validation: {
     MissingOptionalDependencyError,
     XmlValidator
@@ -40,9 +40,10 @@ describe('Serialize.XmlSerialize', function () {
   this.timeout(60000);
 
   [
-    Spec1dot2,
+    Spec1dot5,
+    Spec1dot4,
     Spec1dot3,
-    Spec1dot4
+    Spec1dot2
   ].forEach(spec => describe(`complex with spec v${spec.version}`, () => {
     const normalizerFactory = new XmlNormalizeFactory(spec)