+++ HISTORY.md
@@ -40,6 +40,8 @@ Added functionality regarding [_CycloneDX_ BOM-Link](https://cyclonedx.org/capab
     * Class `JsonValidator` supports _CycloneDX_ Specification-1.5 now ([#505] via [#843])
     * Class `JsonStrictValidator` supports _CycloneDX_ Specification-1.5 now ([#505] via [#843])
     * Class `XmlValidator` supports _CycloneDX_ Specification-1.5 now ([#505] via [#843])
+* Misc
+  * Added functional and integration tests for _CycloneDX_ Specification-1.5 ([#505] via [#843]) 
 
 [#505]: https://github.com/CycloneDX/cyclonedx-javascript-library/issues/505
 [#843]: https://github.com/CycloneDX/cyclonedx-javascript-library/pull/843